code review and removed duplicated methods
